Implemented : admin and user level auth

+++ api/routes/delete.js
@@ -9,19 +9,24 @@ router.post('/',Auth,(req,res,next) => {
         res.status(200).json({message : "you can't delete your own account"});
     }
     else{
-    Users.deleteOne({email:email}).exec()
+    Users.deleteOne(req.userData.type == "user"?{email:email,role:"user"}:{email:email}).exec()
     .then(data =>{
         if(data['deletedCount']>0){
         res.status(200).json({message : email+"'s account has been deleted."});
         }
         else{
-            throw "Email "+email +" not found in the database!"
+            Users.find({email:email})
+            .then(info=>{info.length>0?res.status(403).json({message: "Your not authorized to do this operation"}):res.status(404).json({message: "user not found"})})
+            
+            .catch(err=>{
+                throw err
+            });
         }
     //    res.status(200).json({data : data});
 
         
     }).catch(err =>{
-        res.status(404).json({message: "Error occured while removing the user!",error: err});
+        res.status(500).json({error: err});
     });
 }
 });
